@@ -12,7 +12,7 @@
         "role": "Developer"
     }],
     "require": {
-        "php": ">=5.6.0",
+        "php": ">=7",
       "ext-curl": "*",
       "ext-json": "*"
     },
